Add delegate method to respond to access request
Now the app automatically redirects the user to the page requesting
access to his Dropbox account. If he grants or denies access, the user
is automatically redirected back into the app.

The delegate method:

```objective-c
- (BOOL)application:(UIApplication *)application
            openURL:(NSURL *)url
  sourceApplication:(NSString *)sourceApplication
         annotation:(id)annotation
```

was added to the class `AppDelegate.m`. This method handles the response
the user provides regarding access to his Dropbox. In its current state,
this method does nothing.

A helper method named `appURI` was added the utilities class. It builds
a valid URL based on the URL scheme and URL identifier values defined in
`Info.plist`. This URI is sent as an argument to the OAuth2 parameter
`redirect_uri`.