Add method to request Dropbox user info
Added a method which requests the user's info registered with Dropbox
and prints that info to the console. This method was used as a
proof-of-concept test to see whether the OAuth2 connection was working
as expected. It works!

A caveat to this commit is I had to adjust how credentials are stored.
The property `host` of `NSURL` returns the domain and subdomain
together. Dropbox uses 3 different subdomains in the entire OAuth2
workflow: www, api and api-content. All 3 subdomains use the same access
token, but if we scope the access token to the `host` property, it can't
be seen by other hosts. So, I had to strip out the subdomain and scope
the credentials to the domain.