Save service status
In addition to saving the OAuth2 token in the app's keychain, we also
save the status of the app's relation to the serivce. That is, whether
the app is connected to the service, the last URL that was queried, the
last time a query was made, and so on.

This feature of saving the status persistently in Core Data has been
added.